added dynamic supply
