+++ packages/ui/src/app/dashboard/page.tsx
@@ -2,7 +2,7 @@
 
 import { useContext, useEffect, useState } from 'react';
 import { Button } from '@/components/ui/button';
-import { Plus, Settings, Coins, Trash2, ExternalLink } from 'lucide-react';
+import { Plus, Coins } from 'lucide-react';
 import Link from 'next/link';
 import {
   Card,
@@ -22,6 +22,7 @@ import { TokenDisplay } from '@/types/token';
 import { getAllTokens, getTokenCount } from '@/lib/token/tokenData';
 import { TokenStorage } from '@/lib/token/tokenStorage';
 import { SelectedWalletAccountContext } from '@/context/SelectedWalletAccountContext';
+import { TokenCard } from './components/TokenCard';
 
 export default function DashboardPage() {
   const [selectedWalletAccount] = useContext(SelectedWalletAccountContext);
@@ -59,21 +60,7 @@ function DashboardConnected({ publicKey }: { publicKey: string }) {
     }
   };
 
-  const formatDate = (dateString?: string) => {
-    if (!dateString) return 'Unknown';
-    return new Date(dateString).toLocaleDateString();
-  };
 
-  const getTokenTypeLabel = (type?: string) => {
-    switch (type) {
-      case 'stablecoin':
-        return 'Stablecoin';
-      case 'arcade-token':
-        return 'Arcade Token';
-      default:
-        return type || 'Unknown';
-    }
-  };
 
   if (loading) {
     return (
@@ -194,114 +181,12 @@ function DashboardConnected({ publicKey }: { publicKey: string }) {
 
         <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
           {tokens.map((token, index) => (
-            <Card key={index} className="h-full flex flex-col">
-              <CardHeader>
-                <div className="flex items-center justify-between">
-                  <div className="flex items-center">
-                    <Coins className="h-6 w-6 text-primary mr-3" />
-                    <div>
-                      <CardTitle className="text-lg">
-                        {token.name || `Token ${index + 1}`}
-                      </CardTitle>
-                      <CardDescription>{token.symbol || 'TKN'}</CardDescription>
-                    </div>
-                  </div>
-                  <DropdownMenu>
-                    <DropdownMenuTrigger asChild>
-                      <Button variant="ghost" size="sm">
-                        <Settings className="h-4 w-4" />
-                      </Button>
-                    </DropdownMenuTrigger>
-                    <DropdownMenuContent align="end">
-                      <DropdownMenuItem asChild>
-                        <Link href={`/dashboard/manage/${token.address}`}>
-                          <Settings className="h-4 w-4 mr-2" />
-                          Manage
-                        </Link>
-                      </DropdownMenuItem>
-                      {token.address && (
-                        <DropdownMenuItem asChild>
-                          <a
-                            href={`https://solscan.io/token/${token.address}`}
-                            target="_blank"
-                            rel="noopener noreferrer"
-                          >
-                            <ExternalLink className="h-4 w-4 mr-2" />
-                            View on Solscan
-                          </a>
-                        </DropdownMenuItem>
-                      )}
-                      <DropdownMenuItem
-                        onClick={() => handleDeleteToken(token.address!)}
-                        className="text-red-600"
-                      >
-                        <Trash2 className="h-4 w-4 mr-2" />
-                        Delete from Storage
-                      </DropdownMenuItem>
-                    </DropdownMenuContent>
-                  </DropdownMenu>
-                </div>
-              </CardHeader>
-              <CardContent className="flex-1">
-                <div className="space-y-2 text-sm">
-                  <div className="flex justify-between">
-                    <span className="text-muted-foreground">Type:</span>
-                    <span className="font-medium">
-                      {getTokenTypeLabel(token.type)}
-                    </span>
-                  </div>
-                  <div className="flex justify-between">
-                    <span className="text-muted-foreground">Supply:</span>
-                    <span>{token.supply || '0'}</span>
-                  </div>
-                  <div className="flex justify-between">
-                    <span className="text-muted-foreground">Decimals:</span>
-                    <span>{token.decimals || '6'}</span>
-                  </div>
-                  <div className="flex justify-between">
-                    <span className="text-muted-foreground">Created:</span>
-                    <span>{formatDate(token.createdAt)}</span>
-                  </div>
-                  <div className="flex justify-between">
-                    <span className="text-muted-foreground">Status:</span>
-                    <span className="text-green-600">Active</span>
-                  </div>
-                  {token.extensions && token.extensions.length > 0 && (
-                    <div className="pt-2">
-                      <span className="text-muted-foreground text-xs">
-                        Extensions:
-                      </span>
-                      <div className="flex flex-wrap gap-1 mt-1">
-                        {token.extensions.slice(0, 2).map((ext, idx) => (
-                          <span
-                            key={idx}
-                            className="px-1 py-0.5 bg-blue-100 text-blue-800 rounded text-xs"
-                          >
-                            {ext}
-                          </span>
-                        ))}
-                        {token.extensions.length > 2 && (
-                          <span className="px-1 py-0.5 bg-gray-100 text-gray-600 rounded text-xs">
-                            +{token.extensions.length - 2} more
-                          </span>
-                        )}
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </CardContent>
-              <CardFooter>
-                <Link
-                  href={`/dashboard/manage/${token.address}`}
-                  className="w-full"
-                >
-                  <Button variant="outline" className="w-full">
-                    <Settings className="h-4 w-4 mr-2" />
-                    Manage
-                  </Button>
-                </Link>
-              </CardFooter>
-            </Card>
+            <TokenCard
+              key={index}
+              token={token}
+              index={index}
+              onDelete={handleDeleteToken}
+            />
           ))}
         </div>
       </div>
